refactor(alerts): Move structs into implementing files
