+++ pkg/alerts/conditions.go
@@ -6,6 +6,39 @@ import (
 	"github.com/newrelic/newrelic-client-go/pkg/errors"
 )
 
+// Condition represents a New Relic alert condition.
+// TODO: custom unmarshal entities to ints?
+type Condition struct {
+	PolicyID            int                  `json:"-"`
+	ID                  int                  `json:"id,omitempty"`
+	Type                string               `json:"type,omitempty"`
+	Name                string               `json:"name,omitempty"`
+	Enabled             bool                 `json:"enabled"`
+	Entities            []string             `json:"entities,omitempty"`
+	Metric              string               `json:"metric,omitempty"`
+	RunbookURL          string               `json:"runbook_url,omitempty"`
+	Terms               []ConditionTerm      `json:"terms,omitempty"`
+	UserDefined         ConditionUserDefined `json:"user_defined,omitempty"`
+	Scope               string               `json:"condition_scope,omitempty"`
+	GCMetric            string               `json:"gc_metric,omitempty"`
+	ViolationCloseTimer int                  `json:"violation_close_timer,omitempty"`
+}
+
+// ConditionUserDefined represents user defined metrics for the New Relic alert condition.
+type ConditionUserDefined struct {
+	Metric        string `json:"metric,omitempty"`
+	ValueFunction string `json:"value_function,omitempty"`
+}
+
+// ConditionTerm represents the terms of a New Relic alert condition.
+type ConditionTerm struct {
+	Duration     int     `json:"duration,string,omitempty"`
+	Operator     string  `json:"operator,omitempty"`
+	Priority     string  `json:"priority,omitempty"`
+	Threshold    float64 `json:"threshold,string"`
+	TimeFunction string  `json:"time_function,omitempty"`
+}
+
 // ListConditions returns alert conditions for a specified policy.
 func (alerts *Alerts) ListConditions(policyID int) ([]*Condition, error) {
 	alertConditions := []*Condition{}
