+++ pkg/alerts/nrql_conditions.go
@@ -6,6 +6,28 @@ import (
 	"github.com/newrelic/newrelic-client-go/pkg/errors"
 )
 
+// NrqlCondition represents a New Relic NRQL Alert condition.
+type NrqlCondition struct {
+	Terms               []ConditionTerm `json:"terms,omitempty"`
+	Nrql                NrqlQuery       `json:"nrql,omitempty"`
+	Type                string          `json:"type,omitempty"`
+	Name                string          `json:"name,omitempty"`
+	RunbookURL          string          `json:"runbook_url,omitempty"`
+	ValueFunction       string          `json:"value_function,omitempty"`
+	PolicyID            int             `json:"-"`
+	ID                  int             `json:"id,omitempty"`
+	ViolationCloseTimer int             `json:"violation_time_limit_seconds,omitempty"`
+	ExpectedGroups      int             `json:"expected_groups,omitempty"`
+	IgnoreOverlap       bool            `json:"ignore_overlap,omitempty"`
+	Enabled             bool            `json:"enabled"`
+}
+
+// NrqlQuery represents a NRQL query to use with a NRQL alert condition
+type NrqlQuery struct {
+	Query      string `json:"query,omitempty"`
+	SinceValue string `json:"since_value,omitempty"`
+}
+
 // ListNrqlConditions returns NRQL alert conditions for a specified policy.
 func (alerts *Alerts) ListNrqlConditions(policyID int) ([]*NrqlCondition, error) {
 	conditions := []*NrqlCondition{}
