+++ pkg/alerts/plugins_conditions.go
@@ -6,6 +6,27 @@ import (
 	"github.com/newrelic/newrelic-client-go/pkg/errors"
 )
 
+// PluginsCondition represents an alert condition for New Relic Plugins.
+type PluginsCondition struct {
+	PolicyID          int             `json:"-"`
+	ID                int             `json:"id,omitempty"`
+	Name              string          `json:"name,omitempty"`
+	Enabled           bool            `json:"enabled"`
+	Entities          []string        `json:"entities,omitempty"`
+	Metric            string          `json:"metric,omitempty"`
+	MetricDescription string          `json:"metric_description,omitempty"`
+	RunbookURL        string          `json:"runbook_url,omitempty"`
+	Terms             []ConditionTerm `json:"terms,omitempty"`
+	ValueFunction     string          `json:"value_function,omitempty"`
+	Plugin            AlertPlugin     `json:"plugin,omitempty"`
+}
+
+// AlertPlugin represents a plugin to use with a Plugin alert condition.
+type AlertPlugin struct {
+	ID   string `json:"id,omitempty"`
+	GUID string `json:"guid,omitempty"`
+}
+
 // ListPluginsConditions returns alert conditions for New Relic plugins for a given alert policy.
 func (alerts *Alerts) ListPluginsConditions(policyID int) ([]*PluginsCondition, error) {
 	conditions := []*PluginsCondition{}
