+++ pkg/alerts/synthetics_conditions.go
@@ -6,6 +6,15 @@ import (
 	"github.com/newrelic/newrelic-client-go/pkg/errors"
 )
 
+// SyntheticsCondition represents a New Relic Synthetics alert condition.
+type SyntheticsCondition struct {
+	ID         int    `json:"id,omitempty"`
+	Name       string `json:"name,omitempty"`
+	Enabled    bool   `json:"enabled"`
+	RunbookURL string `json:"runbook_url,omitempty"`
+	MonitorID  string `json:"monitor_id,omitempty"`
+}
+
 // ListSyntheticsConditions returns a list of Synthetics alert conditions for a given policy.
 func (alerts *Alerts) ListSyntheticsConditions(policyID int) ([]*SyntheticsCondition, error) {
 	conditions := []*SyntheticsCondition{}
